core (refactor) rename donthit to hide_from_tree column

+++ core/src/Console/SiteUpdateCommand.php
@@ -132,10 +132,12 @@ class SiteUpdateCommand extends Command
                 $files = explode("\n", file_get_contents($delete_file));
                 foreach ($files as $file) {
                     $file = str_replace('{core}', EVO_CORE_PATH, $file);
-                    if (is_dir($file)) {
-                        SELF::rmdirs($file);
-                    } else {
-                        unlink($file);
+                    if (file_exists($file)) {
+                        if (is_dir($file)) {
+                            removeFolder($file);
+                        } else {
+                            unlink($file);
+                        }
                     }
                 }
             }
